refactor: 계산기 MVC 패턴 적용
